requirements.txt
python-dotenv>=1.0.0
pinecone-client>=3.0.0
lancedb>=0.3.3
yt-dlp>=2023.12.30
moviepy>=1.0.3
pydantic>=2.5.2
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pyyaml>=6.0.1
openai>=1.3.5
langtrace-python-sdk>=0.1.0
aiohttp>=3.9.1
asyncio>=3.4.3
fastapi-cors>=0.0.6
jinja2>=3.1.2
google-cloud-aiplatform>=1.36.0
google-generativeai>=0.3.0
